Add docs for team id
It took me a while to figure out why my projects were not loading and that I had to add a team id to the environment variables. Since I could not find it in the docs I though it would be nice to add it myself. Feel free to change the style and anything else